Correct IndexPack.renamePack to honor standard pack creation rules
C Git always inserts the .pack file into objects/pack before the
.idx file.  This rule ensures that both files are present in the
pack directory before a runtime process tries to open one of them.
This pack insert rule is based upon the pack read rule, which says
a runtime should only open the ".idx" file, and if the ".idx" is
available then the runtime may be able to read the ".pack" too.

We also never attempt to replace an existing pack file.  If the
final pack exists on disk already under this pack name we can
assume it contains the same set of objects.  We never want to
replace known good local objects with objects received from the
network.  This is a simple safety measure to avoid corruption.

+++ org.spearce.jgit/src/org/spearce/jgit/transport/IndexPack.java
@@ -769,21 +769,36 @@ public class IndexPack {
 			oe.copyRawTo(oeBytes, 0);
 			d.update(oeBytes);
 		}
-		ObjectId name = ObjectId.fromRaw(d.digest());
-		File packDir = new File(repo.getObjectsDirectory(), "pack");
-		File finalPack = new File(packDir, "pack-" + name + ".pack");
-		File finalIdx = new File(packDir, "pack-" + name + ".idx");
-		if (!dstIdx.renameTo(finalIdx)) {
-			if (!dstIdx.delete())
-				dstIdx.deleteOnExit();
-			throw new IOException("Cannot rename final pack index");
+
+		final ObjectId name = ObjectId.fromRaw(d.digest());
+		final File packDir = new File(repo.getObjectsDirectory(), "pack");
+		final File finalPack = new File(packDir, "pack-" + name + ".pack");
+		final File finalIdx = new File(packDir, "pack-" + name + ".idx");
+
+		if (finalPack.exists()) {
+			// If the pack is already present we should never replace it.
+			//
+			cleanupTemporaryFiles();
+			return;
 		}
+
 		if (!dstPack.renameTo(finalPack)) {
-			if (!finalIdx.delete())
-				finalIdx.deleteOnExit();
-			if (!dstPack.delete())
-				dstPack.deleteOnExit();
-			throw new IOException("Cannot rename final pack");
+			cleanupTemporaryFiles();
+			throw new IOException("Cannot move pack to " + finalPack);
 		}
+
+		if (!dstIdx.renameTo(finalIdx)) {
+			cleanupTemporaryFiles();
+			if (!finalPack.delete())
+				finalPack.deleteOnExit();
+			throw new IOException("Cannot move index to " + finalIdx);
+		}
+	}
+
+	private void cleanupTemporaryFiles() {
+		if (!dstIdx.delete())
+			dstIdx.deleteOnExit();
+		if (!dstPack.delete())
+			dstPack.deleteOnExit();
 	}
 }
